moved contexts to separate directory
